Reverted binary function changes involving Guarantee
