+++ tests/modules/number_test.cpp
@@ -185,121 +185,3 @@ TEST_F(NumberTest, givenMinBiggerThanMax_shouldThrowInvalidArgument)
     ASSERT_THROW(binary(10, 1), std::invalid_argument);
 }
 
-TEST_F(NumberTest, shouldGenerateBinaryWithGuarantee1)
-{
-    const auto binaryLength = 9;
-
-    GuaranteeMap guarantee{{'0', {3, 7}}, {'1', {2, 7}}};
-    for (int i = 0; i < runCount; ++i)
-    {
-        auto copyGuarantee = guarantee;
-        const auto generatedBinary = binary(std::move(copyGuarantee), binaryLength);
-
-        const auto prefix = generatedBinary.substr(0, 2);
-        const auto binaryNumber = generatedBinary.substr(2);
-
-        ASSERT_EQ(binaryNumber.size(), binaryLength);
-        ASSERT_EQ(prefix, "0b");
-        ASSERT_TRUE(
-            std::ranges::any_of(binaryNumber, [](char binaryNumberCharacter)
-                                { return std::string("01").find(binaryNumberCharacter) != std::string::npos; }));
-        auto count_0 = std::count(binaryNumber.begin(), binaryNumber.end(), '0');
-        auto count_1 = std::count(binaryNumber.begin(), binaryNumber.end(), '1');
-        ASSERT_TRUE(count_0 >= 3 && count_0 <= 7);
-        ASSERT_TRUE(count_1 >= 2 && count_1 <= 7);
-    }
-}
-
-TEST_F(NumberTest, shouldGenerateBinaryWithGuarantee2)
-{
-    const auto binaryLength = 10;
-
-    GuaranteeMap guarantee{{'0', {8, 8}}, {'1', {2, 2}}};
-    for (int i = 0; i < runCount; ++i)
-    {
-        auto copyGuarantee = guarantee;
-        const auto generatedBinary = binary(std::move(copyGuarantee), binaryLength);
-
-        const auto prefix = generatedBinary.substr(0, 2);
-        const auto binaryNumber = generatedBinary.substr(2);
-
-        ASSERT_EQ(binaryNumber.size(), binaryLength);
-        ASSERT_EQ(prefix, "0b");
-        ASSERT_TRUE(
-            std::ranges::any_of(binaryNumber, [](char binaryNumberCharacter)
-                                { return std::string("01").find(binaryNumberCharacter) != std::string::npos; }));
-        auto count_0 = std::count(binaryNumber.begin(), binaryNumber.end(), '0');
-        auto count_1 = std::count(binaryNumber.begin(), binaryNumber.end(), '1');
-        ASSERT_TRUE(count_0 == 8);
-        ASSERT_TRUE(count_1 == 2);
-    }
-}
-
-TEST_F(NumberTest, shouldGenerateBinaryWithGuarantee3)
-{
-    const auto binaryLength = 10;
-
-    GuaranteeMap guarantee{{'0', {10}}};
-    for (int i = 0; i < runCount; ++i)
-    {
-        auto copyGuarantee = guarantee;
-        const auto generatedBinary = binary(std::move(copyGuarantee), binaryLength);
-
-        const auto prefix = generatedBinary.substr(0, 2);
-        const auto binaryNumber = generatedBinary.substr(2);
-
-        ASSERT_EQ(binaryNumber.size(), binaryLength);
-        ASSERT_EQ(prefix, "0b");
-        ASSERT_TRUE(
-            std::ranges::any_of(binaryNumber, [](char binaryNumberCharacter)
-                                { return std::string("01").find(binaryNumberCharacter) != std::string::npos; }));
-        auto count_0 = std::count(binaryNumber.begin(), binaryNumber.end(), '0');
-        ASSERT_TRUE(count_0 == 10);
-    }
-}
-
-TEST_F(NumberTest, shouldGenerateBinaryWithGuarantee4)
-{
-    const auto binaryLength = 10;
-
-    GuaranteeMap guarantee{{'0', {0, 0}}};
-    for (int i = 0; i < runCount; ++i)
-    {
-        auto copyGuarantee = guarantee;
-        const auto generatedBinary = binary(std::move(copyGuarantee), binaryLength);
-
-        const auto prefix = generatedBinary.substr(0, 2);
-        const auto binaryNumber = generatedBinary.substr(2);
-
-        ASSERT_EQ(binaryNumber.size(), binaryLength);
-        ASSERT_EQ(prefix, "0b");
-        ASSERT_TRUE(
-            std::ranges::any_of(binaryNumber, [](char binaryNumberCharacter)
-                                { return std::string("01").find(binaryNumberCharacter) != std::string::npos; }));
-        auto count_0 = std::count(binaryNumber.begin(), binaryNumber.end(), '0');
-        ASSERT_TRUE(count_0 == 0);
-    }
-}
-
-TEST_F(NumberTest, invalidGuaranteeForBinary1)
-{
-    const auto binaryLength = 10;
-
-    GuaranteeMap guarantee{{'0', {6, 10}}, {'1', {6, 10}}};
-    EXPECT_THROW(binary(std::move(guarantee), binaryLength), std::invalid_argument);
-}
-
-TEST_F(NumberTest, invalidGuaranteeForBinary2)
-{
-    const auto binaryLength = 20;
-
-    GuaranteeMap guarantee{{'0', {6, 10}}, {'1', {6, 8}}};
-    EXPECT_THROW(binary(std::move(guarantee), binaryLength), std::invalid_argument);
-}
-
-TEST_F(NumberTest, invalidGuaranteeForBinary3)
-{
-    const auto binaryLength = 10;
-    GuaranteeMap guarantee{{'0', {4}}, {'a', {3}}};
-    EXPECT_THROW(binary(std::move(guarantee), binaryLength), std::invalid_argument);
-}
